♻️ refactor: move dedupeBy to @lobechat/utils package
