+++ packages/utils/src/index.ts
@@ -1,4 +1,5 @@
 export * from './base64';
+export * from './dedupeBy';
 export * from './chunkers';
 export * from './client/cookie';
 export * from './detectChinese';
